+++ Encoding/VCEEnc.vb
@@ -143,6 +143,7 @@ Public Class VCEEnc
 
         Property Custom As New StringParam With {
             .Text = "Custom",
+            .Quotes = QuotesMode.Never,
             .AlwaysOn = True}
 
         Overrides ReadOnly Property Items As List(Of CommandLineParam)
@@ -173,10 +174,14 @@ Public Class VCEEnc
                         New BoolParam With {.Switch = "--vbaq", .Text = "VBAQ"})
                     Add("VUI",
                         New StringParam With {.Switch = "--sar", .Text = "Sample Aspect Ratio", .InitValue = "auto", .Menu = s.ParMenu, .ArgsFunc = AddressOf GetSAR},
+                        New OptionParam With {.Switch = "--videoformat", .Text = "Videoformat", .Options = {"Undefined", "NTSC", "Component", "PAL", "SECAM", "MAC"}},
+                        New OptionParam With {.Switch = "--colormatrix", .Text = "Colormatrix", .Options = {"Undefined", "BT 2020 C", "BT 2020 NC", "BT 470 BG", "BT 709", "FCC", "GBR", "SMPTE 170 M", "SMPTE 240 M", "YCgCo"}},
+                        New OptionParam With {.Switch = "--colorprim", .Text = "Colorprim", .Options = {"Undefined", "BT 2020", "BT 470 BG", "BT 470 M", "BT 709", "Film", "SMPTE 170 M", "SMPTE 240 M"}},
+                        New OptionParam With {.Switch = "--transfer", .Text = "Transfer", .Options = {"Undefined", "ARIB-SRD-B67", "Auto", "BT 1361 E", "BT 2020-10", "BT 2020-12", "BT 470 BG", "BT 470 M", "BT 709", "IEC 61966-2-1", "IEC 61966-2-4", "Linear", "Log 100", "Log 316", "SMPTE 170 M", "SMPTE 240 M", "SMPTE 2084", "SMPTE 428"}},
                         New BoolParam With {.Switch = "--enforce-hrd", .Text = "Enforce HRD compatibility"})
                     Add("Other",
-                        New StringParam With {.Switch = "--chapter", .Text = "Chapters", .Quotes = True, .BrowseFile = True},
-                        New StringParam With {.Switch = "--log", .Text = "Log File", .Quotes = True, .BrowseFile = True},
+                        New StringParam With {.Switch = "--chapter", .Text = "Chapters", .BrowseFile = True},
+                        New StringParam With {.Switch = "--log", .Text = "Log File", .BrowseFile = True},
                         Custom,
                         New OptionParam With {.Switch = "--tier", .Text = "Tier", .Options = {"Main", "High"}, .VisibleFunc = Function() Codec.ValueText = "hevc"},
                         New OptionParam With {.Switch = "--log-level", .Text = "Log Level", .Options = {"Info", "Debug", "Warn", "Error"}},