+++ Encoding/ffmpegEnc.vb
@@ -25,7 +25,6 @@ Public Class ffmpegEnc
         Get
             If ParamsValue Is Nothing Then
                 ParamsValue = New EncoderParams
-
                 ParamsValue.Init(ParamsStore)
             End If
 
@@ -106,6 +105,7 @@ Public Class ffmpegEnc
             proc.WorkingDirectory = p.TempDir
             proc.Package = Package.ffmpeg
             proc.Arguments = args
+            If args.Contains(p.Script.Path) Then proc.Frames = p.Script.GetFrames
             proc.Start()
         End Using
     End Sub
@@ -142,8 +142,11 @@ Public Class ffmpegEnc
             .Switch = "-c:v",
             .Text = "Codec",
             .AlwaysOn = True,
-            .Options = {"x264", "x265", "XviD", "MPEG-4", "Theora", "ProRes", "UT Video", "VP | VP8", "VP | VP9", "Intel | Intel H.264", "Intel | Intel H.265", "Nvidia | Nvidia H.264", "Nvidia | Nvidia H.265"},
-            .Values = {"libx264", "libx265", "libxvid", "mpeg4", "libtheora", "prores", "utvideo", "libvpx", "libvpx-vp9", "h264_qsv", "hevc_qsv", "h264_nvenc", "hevc_nvenc"}}
+            .Options = {"x264", "x265", "AV1", "XviD", "MPEG-4", "Theora", "ProRes", "UT Video",
+                        "VP | VP8", "VP | VP9", "Intel | Intel H.264", "Intel | Intel H.265",
+                        "Nvidia | Nvidia H.264", "Nvidia | Nvidia H.265"},
+            .Values = {"libx264", "libx265", "libaom-av1", "libxvid", "mpeg4", "libtheora", "prores", "utvideo",
+                       "libvpx", "libvpx-vp9", "h264_qsv", "hevc_qsv", "h264_nvenc", "hevc_nvenc"}}
 
         Property Mode As New OptionParam With {
             .Name = "Mode",
@@ -158,6 +161,7 @@ Public Class ffmpegEnc
 
         Property Custom As New StringParam With {
             .Text = "Custom",
+            .Quotes = QuotesMode.Never,
             .AlwaysOn = True}
 
         Overrides ReadOnly Property Items As List(Of CommandLineParam)
@@ -170,6 +174,7 @@ Public Class ffmpegEnc
                         New OptionParam With {.Name = "x264/x265 tune", .Text = "Tune", .Switch = "-tune", .Options = {"None", "Film", "Animation", "Grain", "Stillimage", "Psnr", "Ssim", "Fastdecode", "Zerolatency"}, .VisibleFunc = Function() Codec.OptionText.EqualsAny("x264", "x265")},
                         New OptionParam With {.Switch = "-profile:v", .Text = "Profile", .VisibleFunc = Function() Codec.OptionText = "ProRes", .InitValue = 3, .IntegerValue = True, .Options = {"Proxy", "LT", "Normal", "HQ"}},
                         New OptionParam With {.Switch = "-speed", .Text = "Speed", .AlwaysOn = True, .VisibleFunc = Function() Codec.OptionText.EqualsAny("VP8", "VP9"), .Options = {"6 - Fastest", "5 - Faster", "4 - Fast", "3 - Medium", "2 - Slow", "1 - Slower", "0 - Slowest"}, .Values = {"6", "5", "4", "3", "2", "1", "0"}, .Value = 5},
+                        New OptionParam With {.Switch = "-cpu-used", .Text = "Quality/Speed", .InitValue = 1, .VisibleFunc = Function() Codec.OptionText = "AV1", .IntegerValue = True, .Options = {"0 - Slowest", "1 - Very Slow", "2 - Slower", "3 - Slow", "4 - Medium", "5 - Fast", "6 - Faster", "7 - Very Fast", "8 - Fastest"}},
                         New OptionParam With {.Switch = "-aq-mode", .Text = "AQ Mode", .VisibleFunc = Function() Codec.OptionText = "VP9", .Options = {"Disabled", "0", "1", "2", "3"}, .Values = {"Disabled", "0", "1", "2", "3"}},
                         New OptionParam With {.Name = "h264_nvenc profile", .Switch = "-profile", .Text = "Profile", .Options = {"Baseline", "Main", "High", "High444p"}, .InitValue = 1, .VisibleFunc = Function() Codec.ValueText = "h264_nvenc"},
                         New OptionParam With {.Name = "h264_nvenc preset", .Switch = "-preset", .Text = "Preset", .Options = {"Default", "Slow", "Medium", "Fast", "HP", "HQ", "BD", "LL", "LLHQ", "LLHP", "Lossless", "Losslesshp"}, .InitValue = 2, .VisibleFunc = Function() Codec.ValueText = "h264_nvenc"},
@@ -177,9 +182,8 @@ Public Class ffmpegEnc
                         New OptionParam With {.Name = "h264_nvenc rc", .Switch = "-rc", .Text = "Rate Control", .Options = {"Preset", "Constqp", "VBR", "CBR", "VBR_MinQP", "LL_2Pass_Quality", "LL_2Pass_Size", "VBR_2Pass"}, .VisibleFunc = Function() Codec.ValueText = "h264_nvenc"},
                         New OptionParam With {.Name = "utVideoPred", .Switch = "-pred", .Text = "Prediction", .InitValue = 3, .Options = {"None", "Left", "Gradient", "Median"}, .VisibleFunc = Function() Codec.ValueText = "utvideo"},
                         New OptionParam With {.Name = "utVideoPixFmt", .Switch = "-pix_fmt", .Text = "Pixel Format", .Options = {"YUV420P", "YUV422P", "YUV444P", "RGB24", "RGBA"}, .VisibleFunc = Function() Codec.ValueText = "utvideo"},
-                        New NumParam With {.Name = "Quality", .Text = "Quality", .VisibleFunc = Function() Mode.Value = EncodingMode.Quality AndAlso Not Codec.ValueText.EqualsAny("prores", "utvideo"), .ArgsFunc = AddressOf GetQualityArgs, .Config = {0, 63}},
-                        New NumParam With {.Switch = "-threads", .Text = "Decoding Threads", .Config = {0, 64}},
-                        New NumParam With {.Name = "VPX enc threads", .Switch = "-threads", .Text = "Encoding Threads", .VisibleFunc = Function() Codec.OptionText.EqualsAny("VP8", "VP9"), .Value = 8, .DefaultValue = -1},
+                        New NumParam With {.Name = "Quality", .Text = "Quality", .Init = -1, .VisibleFunc = Function() Mode.Value = EncodingMode.Quality AndAlso Not Codec.ValueText.EqualsAny("prores", "utvideo"), .ArgsFunc = AddressOf GetQualityArgs, .Config = {-1, 63}},
+                        New NumParam With {.Switch = "-threads", .Text = "Threads", .Config = {0, 64}},
                         New NumParam With {.Switch = "-tile-columns", .Text = "Tile Columns", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 6, .DefaultValue = -1},
                         New NumParam With {.Switch = "-frame-parallel", .Text = "Frame Parallel", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 1, .DefaultValue = -1},
                         New NumParam With {.Switch = "-auto-alt-ref", .Text = "Auto Alt Ref", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 1, .DefaultValue = -1},
@@ -228,7 +232,13 @@ Public Class ffmpegEnc
                     ret += $" -b:v {p.VideoBitrate}k"
             End Select
 
-            If Codec.OptionText = "XviD" Then ret += " -tag:v xvid"
+            Select Case Codec.OptionText
+                Case "XviD"
+                    ret += " -tag:v xvid"
+                Case "AV1"
+                    ret += " -strict experimental"
+            End Select
+
             Dim targetPath As String
 
             If Mode.OptionText = "Two Pass" AndAlso pass = 1 Then
@@ -257,7 +267,7 @@ Public Class ffmpegEnc
                 If param.Value <> param.DefaultValue Then
                     If Codec.OptionText.EqualsAny("VP8", "VP9") Then
                         Return "-crf " & param.Value & " -b:v 0"
-                    ElseIf Codec.OptionText.EqualsAny("x264", "x265") Then
+                    ElseIf Codec.OptionText.EqualsAny("x264", "x265", "AV1") Then
                         Return "-crf " & param.Value
                     Else
                         Return "-q:v " & param.Value