+++ Encoding/x264Enc.vb
@@ -235,6 +235,7 @@ Public Class x264Params
 
     Property Custom As New StringParam With {
         .Text = "Custom",
+        .Quotes = QuotesMode.Never,
         .AlwaysOn = True,
         .InitAction = Sub(tb)
                           tb.Edit.Expand = True
@@ -245,6 +246,7 @@ Public Class x264Params
 
     Property CustomFirstPass As New StringParam With {
         .Text = "Custom" + BR + "First Pass",
+        .Quotes = QuotesMode.Never,
         .InitAction = Sub(tb)
                           tb.Edit.Expand = True
                           tb.Edit.TextBox.Multiline = True
@@ -254,6 +256,7 @@ Public Class x264Params
 
     Property CustomSecondPass As New StringParam With {
         .Text = "Custom" + BR + "Second Pass",
+        .Quotes = QuotesMode.Never,
         .InitAction = Sub(tb)
                           tb.Edit.Expand = True
                           tb.Edit.TextBox.Multiline = True